long policy
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@5716 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/export/kernel.h
@@ -78,7 +78,7 @@ extern void queue_wait_w_tmo(struct event_queue *q, struct event *ev, int ticks)
 extern void queue_post(struct event_queue *q, long id, void *data);
 extern bool queue_empty(const struct event_queue* q);
 void queue_clear(struct event_queue* q);
-extern int queue_broadcast(int id, void *data);
+extern int queue_broadcast(long id, void *data);
 
 extern void mutex_init(struct mutex *m);
 extern void mutex_lock(struct mutex *m);
